Add Base.transaction_display
This commit adds Base.transaction_display, which should be either None
or an instance of dnf.yum.rpmtrans.TransactionDisplay (or a subclass
thereof).

If Base.do_transaction() is called without a 'display' argument,
it will use self.transaction_display as the default value.

It defaults to None. BaseCli sets it to output.CliTransactionDisplay().

Moving it to an attribute allows Plugins and/or Commands to add their
own TransactionDisplay - for example, fedup or offline-updates could use
this to send status info to plymouth during the upgrade transaction.